fix: healthcheck card usage in application
