+++ view/packages/hooks/applications/use_deployment_health_chart.tsx
@@ -80,7 +80,7 @@ export function useDeploymentHealthChart({
 
   const customHeader = useMemo(
     () => (
-      <div className="flex items-center justify-between">
+      <div className="flex items-center justify-between w-full">
         <span>{t('selfHost.monitoring.chart.title')}</span>
         {headerActions}
       </div>
